Fboemer/more ops (#29)
* Added slice, reshape, broadcast, concat, one-hot ops
* Parallel dot currently needs to be hard-coded for which axis to parallelize over
* Switched to debug version of SEAL that allows multiplication by 0
* Parallelizes noise budget checking, dot product, and tensor write
* TestHEBackend.tf_mnist_exp_const currently uses ReLu op, and therefore fails.

+++ test/model/mnist_mlp_const_5_inputs_int.js
@@ -3,7 +3,7 @@
         "name": "Function_0",
         "ops": [
             {
-                "element_type": "float",
+                "element_type": "int64_t",
                 "inputs": [],
                 "name": "Parameter_0",
                 "op": "Parameter",
@@ -16,7 +16,7 @@
                 ]
             },
             {
-                "element_type": "float",
+                "element_type": "int64_t",
                 "inputs": [],
                 "name": "Constant_3",
                 "op": "Constant",
@@ -3951,7 +3951,7 @@
                 ]
             },
             {
-                "element_type": "float",
+                "element_type": "int64_t",
                 "inputs": [],
                 "name": "Constant_2",
                 "op": "Constant",
@@ -82366,7 +82366,7 @@
                 ]
             },
             {
-                "element_type": "float",
+                "element_type": "int64_t",
                 "inputs": [],
                 "name": "Constant_1",
                 "op": "Constant",
